cleaned app.js and added style to budget value display
