+++ app/assets/stylesheets/components/_index.scss
@@ -9,3 +9,5 @@
 @import "roulette";
 @import "filters";
 @import "friend-list";
+@import "new-event-form";
+
