+++ app/assets/stylesheets/components/_new-event-form.scss
@@ -1,0 +1,11 @@
+#event_budget_value_display {
+  justify-content: center;
+  font-size: 70px;
+  font-weight: bold;
+}
+
+#value_per_person {
+  text-align: center;
+  font-size: 70px;
+  font-weight: bold;
+}
