+++ app/javascript/components/event_form.js
@@ -117,4 +117,16 @@ formListen();
 
 };
 
-export { eventRest };
+const sliderBudgetValue = () => {
+  const refreshEventBudgetValue = () => {
+    document.querySelector('#event_budget_value').innerHTML = event.target.value;
+  };
+
+  const slider = document.querySelector('#event_budget');
+  if (slider) {
+    document.querySelector('#event_budget_value').innerHTML = slider.value;
+    slider.addEventListener('change', refreshEventBudgetValue);
+  }
+}
+
+export { eventRest, sliderBudgetValue };
