+++ app/javascript/packs/application.js
@@ -3,6 +3,8 @@ import "bootstrap";
 // import { formListen } from '../components/event_form';
 import { eventRest } from '../components/event_form';
 import { tagFunctionality } from '../components/tag_choice';
+import { sliderBudgetValue } from '../components/event_form';
+
 
 import { editionRest } from '../components/roulette';
 // import { spinRoulette } from '../components/roulette';
@@ -10,7 +12,7 @@ import { editionRest } from '../components/roulette';
 // // NEW EVENT FORM
 const tagForm = document.querySelector("#new_event");
 if (tagForm) {
-  console.log("new event")
+  sliderBudgetValue();
   tagFunctionality();
   eventRest();
 };
@@ -20,13 +22,3 @@ const roulettePage = document.querySelector('#wheel');
 if (roulettePage) {
   editionRest();
 };
-
-const refreshEventBudgetValue = () => {
-  document.querySelector('#event_budget_value').innerHTML = event.target.value;
-};
-
-const slider = document.querySelector('#event_budget');
-if (slider) {
-  document.querySelector('#event_budget_value').innerHTML = slider.value;
-  slider.addEventListener('change', refreshEventBudgetValue);
-}
