+++ app/views/events/new.html.erb
@@ -32,8 +32,13 @@
 
   <fieldset class="field-slide msf_hide">
     <div class="event-slide d-flex flex-column justify-content-between">
-      <%= f.range_field :budget, min: "0", max: "600", label: 'How much do you want to spend?' %>
-      <div id="event_budget_value"></div>
+      <p>How much do you want to spend?</p>
+      <%= f.range_field :budget, min: "0", max: "150" %>
+      <div class="d-flex" id="event_budget_value_display">
+        <div id="event_budget_value"></div><span>€</span>
+      </div>
+      <div id="value_per_person">per person</div>
+      <p>Hint: Use keyboard arrows if you prefer 😉</p>
       <div>
         <div class="d-flex justify-content-between">
           <input type="button" name="back" value="Back" id="button-back3" class="btn btn-success">
